### 2020-12-11 #416 @mikemahoney218
terrainr: Retrieve Data from the USGS National Map and Transform it for 3D Landscape Visualizations
Submitting Author: Michael Mahoney (@mikemahoney218)
Repository: https://github.com/mikemahoney218/terrainr
Version submitted: 0.2.1
Version accepted: Feb. 19, 2021

---

```
Package: terrainr
Title: Retrieve Data from the 'USGS' National Map and Transform it for '3D' Landscape Visualizations
Version: 0.2.0
    person(given = "Michael",
           family = "Mahoney",
           email = "mike.mahoney.218@gmail.com",
           comment = c(ORCID = "0000-0003-2402-304X"))
Description: Functions for the retrieval and manipulation of 'USGS' National Map
    data ('<https://viewer.nationalmap.gov/services/>'), enabling users to
    download elevation and image data for areas of interest. Functions are also
    provided to transform these data sources into formats that can be used to
    create 3D elevation models in the Unity rendering engine.
    base64enc,
    httr,
    raster,
    magick (>= 2.5.0),
    gdalUtils,
    png,
    gdalUtilities,
    sf,
    rlang
    covr,
    progressr,
    progress,
    jpeg,
    tiff
Config/testthat/parallel: true
```

`terrainr` enables users to retrieve geospatial data from the USGS National Map family of APIs and process it into both generic rasters for use in analysis and a specialized format which may be imported into the Unity rendering engine.

The data retrieval functions within terrainr are designed to be useful to any researcher interested in geospatial data for areas within the United States, particularly digital elevation models and orthoimagery. The data processing aspect of terrainr is designed to enable more users to use game engines as GIS, producing high-resolution 3D models from spatial data.

I am only aware of two packages which mention the USGS National Map API:

* [quickmapr](https://github.com/jhollist/quickmapr) provides the ability to download a topographic basemap for use in plotting. [The URL used](https://github.com/jhollist/quickmapr/blob/6824134188d28d9c0fb584dd0922bfa26811a1f1/R/qmap.R#L229) does not overlap with the APIs included in this package, and I believe is not actually part of the National Map at all (as images from the National Map are released into the public domain, while the referenced URL indicates that downloaded data is copyright the National Geographic Society). This package was additionally last updated in 2018, and [an entry in NEWS.md](https://github.com/jhollist/quickmapr/blob/master/NEWS.md#quickmapr-v012) indicates the overlapping functionality may have been destined for deprecation.
* A prior package of mine, [spacey](https://github.com/mikemahoney218/spacey), provides access to the 3DEP Digital Elevation Model, one of the APIs included in this package. Future development of `spacey` will move to using `terrainr` for data retrieval.

The [elevatr](https://github.com/jhollist/elevatr/) package also returns elevation data from the USGS Elevation Point Query Service and AWS Terrain Tiles, which are similar -- but not the same -- sources as used in this package.

This package provides access to 9 API endpoints which I do not believe are incorporated in any other R package. Transforming tiles to be imported into Unity as heightmaps depends on [a newly-contributed feature to the magick package](https://github.com/ropensci/magick/commit/2183f08983a8a0ef1b8e32ec28b7607e94a232b0), which leads me to believe this package is the first to provide this functionality.

N/A

N/A
