### 2021-01-05 #416 @ldecicco-USGS
I had a few people respond that if I didn't find someone by the new year, they'd volunteer, so I expect a 2nd reviewer any day. In the meantime, I ran a check and `goodpractice::gp()`.  The check was fine.

The `gp()` left this message:
```
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck
── GP terrainr ──────────────────────────────────────────────────────

It is good practice to

  ✖ write unit tests for all functions, and all
    package code in general. 43% of code lines are covered by
    test cases.

    R/combine_overlays.R:119:NA
    R/georeference_overlay.R:76:NA
    R/get_bbox.R:48:NA
    R/get_bbox.R:49:NA
    R/get_tiles.R:133:NA
    ... and 458 more lines

  ✖ write short and simple functions. These
    functions have high cyclomatic complexity:get_tiles (55).
─────────────────────────────────────────────────────────────────────
```
Which is not bad! If you want to work on the code coverage, that's always good.
