### 2021-01-05 #416 @mikemahoney218
Fantastic! Thank you so much for the update :)

One thing to mention -- a lot of the package tests are set to skip on CRAN, as they either take too long, require internet access, or require GDAL. They should still be able to run locally, however (and on CI) with `Sys.setenv(NOT_CRAN='true')` -- when I do so, I get the following output:

```
> Sys.setenv(NOT_CRAN='true'); goodpractice::gp()

Preparing: covr
Preparing: cyclocomp
<snip r cmd check>
Preparing: description
Preparing: lintr
Preparing: namespace
Preparing: rcmdcheck
── GP terrainr ──────────────────────────────────────────────────────
It is good practice to

    R/combine_overlays.R:119:NA
    R/get_bbox.R:48:NA
    R/get_bbox.R:49:NA
    R/hit_api.R:160:NA
    R/hit_api.R:161:NA
    ... and 26 more lines

  ✖ write short and simple functions. These functions have high cyclomatic complexity:get_tiles (55).
─────────────────────────────────────────────────────────────────────
```

Thanks again for the update! :smile:
