### 2021-01-18 #416 @mikemahoney218
Hi all! Just wanted to note here that I've changed the main branch from `master` to `main` in line with [GitHub's guidance](https://github.com/github/renaming). There shouldn't be any code changes (well, a single call got restyled) or anything to interrupt review. Sorry if there are any issues associated with the change!
