### 2021-01-18 cd064ff
Schedule tests and make hit_api smaller
