+++ .github/workflows/check-ubuntu-oldrel.yaml
@@ -2,11 +2,11 @@ on:
   push:
     branches:
       - main
-      - master
   pull_request:
     branches:
       - main
-      - master
+  schedule: 
+    - cron: 10 7 * * 1,3,5
 
 name: R-CMD-check-Ubuntu-oldrel
 
