+++ .github/workflows/check-ubuntu-devel.yaml
@@ -2,11 +2,11 @@ on:
   push:
     branches:
       - main
-      - master
   pull_request:
     branches:
       - main
-      - master
+  schedule: 
+    - cron: 30 7 * * 1,3,5
 
 name: R-CMD-check-Ubuntu-devel
 
