### 2021-01-18 #416 @mikemahoney218
I believe I've spoken too soon! It appears that the `transportation` API has moved servers, so version 0.2.0 no longer works for that endpoint. I've pushed a fix to the development branch ([here's the relevant diff](https://github.com/mikemahoney218/terrainr/commit/16b14e1a5796f57bc7c460570f35c7298fba2c65)) -- it was only a two-line change to handle `transportation` in the same way as DEMs and orthoimagery.

I apologize for needing to make this update while the package is under review! The version originally submitted still exists as [release v 0.2.0](https://github.com/mikemahoney218/terrainr/tree/32e92b9c3f7c9e57fb92f0df5472a60e6199b22e). Does it make sense to update the version under review to the version with this fix (would become version 0.2.1), or should I wait and roll this update in with any revisions requested after review?
