### 2021-01-18 ebae2b0
Style and update NEWS for 0.2.1
