### 2021-01-18 #416 @mikemahoney218
Great! I've cut a [release of 0.2.1](https://github.com/mikemahoney218/terrainr/releases/tag/v0.2.1) and split it into a [release branch](https://github.com/mikemahoney218/terrainr/tree/0.2.1), both of which should hopefully remain even with `main` until reviewer comments come in.

Thanks!
