+++ .github/workflows/check-ubuntu-oldrel.yaml
@@ -6,7 +6,7 @@ on:
     branches:
       - main
   schedule: 
-    - cron: 10 7 * * 1,3,5
+    - cron: 10 12 * * 1,3,5
 
 name: R-CMD-check-Ubuntu-oldrel
 
