+++ .github/workflows/check-ubuntu-release.yaml
@@ -6,7 +6,7 @@ on:
     branches:
       - main
   schedule: 
-    - cron: 0 8 * * 1,3,5
+    - cron: 0 13 * * 1,3,5
 
 name: R-CMD-check-Ubuntu-release
 
