+++ .github/workflows/check-windows.yaml
@@ -6,7 +6,7 @@ on:
     branches:
       - main
   schedule: 
-    - cron: 0 7 * * 1,3,5
+    - cron: 0 12 * * 1,3,5
 
 name: R-CMD-check-Windows
 
