+++ .github/workflows/check-mac.yaml
@@ -6,7 +6,7 @@ on:
     branches:
       - main
   schedule: 
-    - cron: 45 7 * * 1,3,5
+    - cron: 45 12 * * 1,3,5
 
 name: R-CMD-check-MacOS
 
