+++ .github/workflows/check-ubuntu-devel.yaml
@@ -6,7 +6,7 @@ on:
     branches:
       - main
   schedule: 
-    - cron: 30 7 * * 1,3,5
+    - cron: 30 12 * * 1,3,5
 
 name: R-CMD-check-Ubuntu-devel
 
