### 2021-02-02 54ad9fb
Remove manual href categorization
