+++ .github/workflows/lint.yaml
@@ -1,12 +1,12 @@
 on:
   push:
     branches:
-      - main
-      - master
+      - '*'
+      - '!gh-pages'
   pull_request:
     branches:
-      - main
-      - master
+      - '*'
+      - '!gh-pages'
 
 name: lint
 
