+++ .github/workflows/check-mac.yaml
@@ -1,10 +1,12 @@
 on:
   push:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   pull_request:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   schedule: 
     - cron: 45 12 * * 1,3,5
 
