+++ .github/workflows/check-ubuntu-release.yaml
@@ -1,12 +1,14 @@
 on:
   push:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   pull_request:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   schedule: 
-    - cron: 0 13 * * 1,3,5
+    - cron: 45 12 * * 1,3,5
 
 name: R-CMD-check-Ubuntu-release
 
