+++ .github/workflows/test-coverage.yaml
@@ -1,10 +1,12 @@
 on:
   push:
     branches:
-      - master
+      - '*'
+      - '!gh-pages'
   pull_request:
     branches:
-      - master
+      - '*'
+      - '!gh-pages'
 
 name: test-coverage
 
