+++ .github/workflows/check-windows.yaml
@@ -1,12 +1,15 @@
 on:
   push:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   pull_request:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   schedule: 
-    - cron: 0 12 * * 1,3,5
+    - cron: 45 12 * * 1,3,5
+
 
 name: R-CMD-check-Windows
 
