+++ .github/workflows/check-ubuntu-devel.yaml
@@ -1,12 +1,14 @@
 on:
   push:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   pull_request:
     branches:
-      - main
+      - '*'
+      - '!gh-pages'
   schedule: 
-    - cron: 30 12 * * 1,3,5
+    - cron: 45 12 * * 1,3,5
 
 name: R-CMD-check-Ubuntu-devel
 
