### 2021-02-02 ff40295
Restyle file, bump version, add NEWS
