### 2021-02-05 d15a8b6
Small cleanup
