### 2021-02-10 #416 @mikejohnson51
Hi @mikemahoney218 @ldecicco-USGS ,

Please find my review of `terrainr` attached:

Overall this is a very nice package. I want to emphasize this as, by nature, the following comments are more critique then complimentary. The package functions well, is stable and installed without problems.

I did not get to test any of the Unity functionality since I don't have it. Other then that here are some key points:

## Key Points:

- Really nice website and great documentation.

- All functions and examples worked (except one noted below)

- I think there is some nice utility here but there is also some cross over with others packages like `FedData`, `elevatr`.

- Following this, other then elevation, this is not a "spatial data" retrieval package as it returns PNGs that are more similar to map tiles then to spatial data. Thus the functionality is also similar to `rosm`.

- I cannot think of many cases for wanting the PNG files in R. That **certainly** does not mean they are not out there but I think more clarity about what the package is returning (maptiles) and there use would help it gain traction.

- The major issue I have with the package as it stands is the package specific classes for common spatial structures like coordinates and bounding boxes. They add more confusion and complexity to things that are very well served by the `sf` package (which is already a dependency). Utilizing the `sf` utilities would not only allow for more flexible user input, but would drastically reduce the code in this package making it much slimmer and easier to maintain.

- Aside from the "new" spatial classes, there is no need to reinvent the wheel for bbox buffering. Instead focus on the input/output coordinate reference systems and use `st_buffer.` Same thing for finding a centroid (`st_centroid`) and working between units (`units::set_units()`)

- I highly recommend using the `sf` utilities for GDAL rather then bringing in other dependencies like `gdalUtilities` and `gdalUtils`. Some examples are included here for reducing the clever, but not needed work arounds offered

- Use `httr::RETRY` to handle much of your error catching and `httr::write_disk()` to avoid having to read images in as raw bytes/bits.  This will save a large amount of code as well.

- With all of these I think you can drop: `base64enc`,  `gdalUtils`, `methods`, `gdalUtilities`, `gdalUtils` and `rlang` from your dependencies and many of your function can be slimmed down and/or removed. The result (in my option) will be a more lightweight package with higher utility, and more easy integration in spatial workflows.

## Branding

- It's a weird point but one that I think is important. `terrainr` as a name sells your package short while also being misleading. Terrain is one of 9 endpoints offered but is also anomalous in that it is the only "spatial" data available in your package (in the traditional vector/raster sense).

- Arguably terrain is also the best served of the resources in the R ecosystem with `elevatr` and `FedData`. `Elevatr` on the surface is a better choice for getting elevation data simply because it offers multiple resolutions and the Mapzen server is quicker then the ArcGIS server.

- My expectation of the package coming in given the description of "retrieve geospatial data" was that for resources like the NHD, NHDHighRes, WBD, and transportation, I would be getting the vector data. Particularly since these are offered by the National Map. Therefore, it is very important to be upfront that the package is offering endpoints to (effectively) mosaicked web tiles (with one spatial data source).

- All of this is not to discount the utility of your package because I think it is one of the few to offer access to these resources which makes it very unique! But the functionality needs to be explicit because, as is, someone coming in with the expectation that they will be able to query spatial data will be disappointed, while those looking for map tile data will likely pass your package by.

- One **huge** opportunity for your package, that I think you already have worked out, is how to integrated these resources with ggplot as a base-map. It is a constant frustration of mine (and many I know) to get nice base maps for ggplot.

- Yes, `tmap` has some utilities for this and `ggmap` is out there (I have always struggled with it), there is room here for growth. If you could add a `geom_terrainr()` that played nicely with `ggplot`, I think that would be immediately useful to a many people.

# Code Review:

Below are my notes as I went through the files and package:

## Install

```{r}
devtools::install_github("mikemahoney218/terrainr")
library(terrainr)
```

Great! 👍

##  `add_bbox_buffer.R`

There are a lot of functions in this package that recreate `sf` functionality. These could be simplified making the code much easier to maintain. While the implementation of these methods is obviously well considered, it is fragile compared to the GDAL/GEOS/PROJ backend supporting `sf`. For this package to attract wide use I think the package specific classes/structures must be removed For example, `add_bbox_buffer`:

Could be simplified to something like:

```{r}
library(sf)
bb = st_as_sfc(st_bbox(c(ymin = 44.17609, ymax = 44.04905,
                         xmin = -74.01188, xmax = -73.83493), crs = 4326))
tmp = st_transform(bb, 5070) # EPSG ensures meters
tmp = st_buffer(tmp, 10,
                joinStyle = 'MITRE',
                endCapStyle = "SQUARE",
                mitreLimit = 2)

out = st_transform(tmp, 4326) # Put back to Lat Long
```

- It is certainly a shameless self plug but check out the AOI/climateR workflows. AOI allow users to (1) flexably generate AOIs and (2) climateR simply takes a spatial object and can identify the underlying climate grids:

  - Build AOIs: https://github.com/mikejohnson51/AOI
  - Use AOIs to get climate grids: https://github.com/mikejohnson51/climateR

- The new nhdplusTools 'get' family (get_nhdplus(), get_huc8(), get_waterbodies()) also use this logic:

  - Example: https://github.com/USGS-R/nhdplusTools/blob/master/R/get_nhdplus.R

- Think about this kind of workflow for more generic queries.

- `elevatr` also offers a nice way that is more generic:

https://github.com/jhollist/elevatr/blob/4c45d548231400284f72b92b42dff6d9fd9a0928/R/get_elev_raster.R#L9

## `classes.R`

I think your toolset has a lot of value outside the `terrainr` environment, but it becomes very limited with the imposed classes for things like coordinate sets and bounding boxes which already have well defined and common class structures that play nicely in the R spatial environment. I would like to see all the specific classes removed from the package or else a strong justification for what they add. I am aware this will take some considerable rewrite in the package but think it is well worth it.

# `get_bbox.R`

Again `sf::st_bbox` does everything needed. No need to write unique methods to extract bbox's from common spatial classes. This removes your dependency of `rlang` too.

For example:

```{r}
st_bbox(c(ymin = 44.04905, ymax = 44.17609,
                         xmin = -74.01188, xmax = -73.83493), crs = 4326)

st_bbox(raster::raster(ncol=100, nrow=100))
```

# `get_bbox_centroid`

Same thing. `st_centroid()` does what you need. No need for specific classes and logic. For example:

```{r}
get_bbox_centroid(list(
     c(lat = 44.04905, lng = -74.01188),
     c(lat = 44.17609, lng = -73.83493)))

# Equals

bb = st_as_sfc(st_bbox(c(ymin = 44.04905, ymax = 44.17609,
                         xmin = -74.01188, xmax = -73.83493), crs = 4326))

st_centroid(bb)
```

# `get_tiles.R`

- This function, even for the example, takes a while to run, it would be nice to have some messaging along the way to give confidence it is going if `progressr` is not installed. Consider using `httr::progress()` in `hit_national_map_api`. More on that below.

- Its nice to be able to provide a `tmpfile` prefix, but would be equally nice to be able to set where the file is downloaded to (other then the tmp directory). This will come up again below with the GDAL discussion.

- find a way to avoid the `3` in the beginning of your list name for 3DEP, maybe elevation3DEP?

# `hit_national_map_api.R`

- I suggest using `httr::RETRY("GET", ...)`  rather then a tryCatch and counters, this will greatly simplify your logic. For example:

```{r}
#Set UP
url = httr::parse_url("https://elevation.nationalmap.gov/arcgis/rest/services/3DEPElevation/ImageServer/exportImage")

query_arg <- list(
      bboxSR = 4326,
      imageSR = 4326,
      size = paste(8000, 8000, sep = ","),
      format = "tiff",
      pixelType = "F32",
      noDataInterpretation = "esriNoDataMatchAny",
      interpolation = "+RSP_BilinearInterpolation",
      f = "json"
    )

##
first_corner <- bbox@bl
second_corner <- bbox@tr

bbox_arg <- list(bbox = paste(
    min(first_corner@lng, second_corner@lng),
    min(first_corner@lat, second_corner@lat),
    max(second_corner@lng, first_corner@lng),
    max(second_corner@lat, first_corner@lat),
    sep = ","
  ))

####======
#^all the above could be replaces with paste(st_bbox(...), collapse = ",")
````

You have a lot of logic wrapping your `httr` calls which is essentially:

```{r}
res <- httr::GET(url, query = c(bbox_arg, query_arg))
content = httr::content(res, type = "application/json")
```

An identical, but way to build in some of that error catching and retrying would be to change it to:

```{r, eval = F}
res2 <- httr::RETRY("GET", url,
                   query = c(bbox_arg, query_arg),
                   times = 10, pause_cap = 10)
```

You could  use this approach again to get the images and write them to disk, instead of `httr::content(..., "raw")`. For example:

```{r, eval = FALSE}
tmpfile <-tempfile()
content <-httr::content(res2, type = "application/json")

httr::RETRY("GET", content$href,
                   httr::progress(),
                   httr::write_disk(tmpfile, overwrite = TRUE),
                   times = 10, pause_cap = 10)

raster::raster(tmpfile) %>% raster::plot()
```

# `merge_rasters.R`

The GDAL stuff you have here is (reasonably!!) way more than needed. One big reason to avoid `gdalUtils` in a package is because it requires an instance of GDAL that is different than the version that comes with `sf`. For many users this can be an issue.
Here is an brief solution to these issues for much less code and much more utility:

The example you have only gives one raster which I assume wouldn't need to be merged? This should probably be changed, but I pulled an example from one of your other functions:

```{r}
simulated_data <- data.frame(
  id = seq(1, 100, 1),
  lat = runif(100, 44.04905, 44.17609),
  lng = runif(100, -74.01188, -73.83493)
)

bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
bbox <- add_bbox_buffer(bbox, 100)
img_files = get_tiles(bbox, tempfile())
```

So the goal is to get this to a single raster right? If so try:

```{r, eval = FALSE}
target_prj = st_crs(5070)$proj4string
method = "near"
destfile1 = tempfile()

sf::gdal_utils(util = "warp",
                 source = unlist(img_files),
                 destination = destfile1,
                 options = c("-t_srs", as.character(target_prj),
                             "-r", method))

raster::raster(destfile1)

# crs        : +proj=aea +lat_0=23 +lon_0=-96 +lat_1=29.5 +lat_2=45.5 +x_0=0 +y_0=0 +datum=NAD83 +units=m +no_defs
```

The cool thing about above is that it is short-and-sweet, only relies on `sf` and it offers you a direct path to allow users to specify the output CRS and resampling method (exposing `target_prj`,  `method`, `destfile1` as function parameters.)

Another reason to keep your bbox objects in standard classes is that you can use then to quickly crop the merged raster in the same call by first writing the bbox object to a shp. So instead of the above you could do:

```{r}
# From the same coordinates make a 'sf' bbox of your points:
bbox = st_as_sf(simulated_data, coords = c("lng", "lat"), crs = 4326) %>%
  st_bbox() %>%
  st_as_sfc() %>%
  st_as_sf()

# Write temporary shp
tmp_shp = tempfile(fileext = ".shp")
write_sf(bbox, dsn = tmp_shp)

destfile2 = tempfile()

sf::gdal_utils(util = "warp",
                 source = unlist(img_files),
                 destination = destfile2,
                 options = c("-t_srs", as.character(target_prj),
                             "-r", method,
                             '-cutline', tmp_shp,
                             '-crop_to_cutline'))

raster::raster(destfile2)

# names      : file220c7958114b
```

# `raster_to_raw_tiles.R`

Use `sf::gdal_utils("translate", ...)` here instead of gdalUtils/gdalUtilities for the same reasons as above.

It may be worth your time to look at this: https://github.com/ropensci/tiler

### georeference

The given example does not run because the arguments are in the wrong order, also it doesn't seem to need alignment? The input and output are identical.

No need for TIFF package (I think) since `raster` reads tiffs already (raster will also read PNG and JPEG, just into RGBA bands)

# `utils.R`

Check out the `units` package to handle all these unit conversions, but really, the unit given should match the CRS of the input data. Because of this, I might consider being more strict about the units you allow for a buffering call.

Notes:

  - Readme - you have an 'elev' hanging at the end of the 2nd 'plot' line

Again, there's a lot of good stuff here, I hope this is useful, and feel free to email for any clarification/follow up.

Mike
