+++ README.md
@@ -43,15 +43,6 @@ output_tiles <- get_tiles(bbox = bbox,
                           services = c("elevation", "ortho"))
 ```
 
-``` r
-# output_tiles is now a list of two vectors pointing to the elevation and 
-# orthoimagery tiles we just downloaded -- here we're displaying the first
-# of the ortho tiles
-raster::plot(raster::raster(output_tiles[[2]][[1]]))
-```
-
-<img src="man/figures/naip.png" width="100%" />
-
 Once downloaded, these images are in standard GeoTIFF or PNG formats and
 can be used as expected with other utilities:
 
@@ -70,10 +61,8 @@ VR:
 merged_dem <- tempfile(fileext = ".tif")
 merged_ortho <- tempfile(fileext = ".tif")
 # we can call these vectors by name instead of position, too
-merge_rasters(output_tiles$`3DEPElevation`, 
-              merged_dem, 
-              output_tiles$USGSNAIPPlus, 
-              merged_ortho)
+merge_rasters(output_tiles[["elevation"]], merged_dem)
+merge_rasters(output_tiles[["ortho"]], merged_ortho)
 
 mapply(function(x, y) raster_to_raw_tiles(input_file = x, 
                                           output_prefix = tempfile(), 
@@ -104,39 +93,39 @@ Unity.](https://mikemahoney218.github.io/terrainr/articles/unity_instructions.ht
 The following datasets can currently be downloaded using `get_tiles` or
 `hit_national_map_api`:
 
-  - [3DEPElevation](https://elevation.nationalmap.gov/arcgis/rest/services/3DEPElevation/ImageServer):
+-   [3DEPElevation](https://elevation.nationalmap.gov/arcgis/rest/services/3DEPElevation/ImageServer):
     The USGS 3D Elevation Program (3DEP) Bare Earth DEM.
-  - [USGSNAIPPlus](https://services.nationalmap.gov/arcgis/rest/services/USGSNAIPPlus/MapServer):
+-   [USGSNAIPPlus](https://services.nationalmap.gov/arcgis/rest/services/USGSNAIPPlus/MapServer):
     National Agriculture Imagery Program (NAIP) and high resolution
     orthoimagery (HRO).
-  - [nhd](https://hydro.nationalmap.gov/arcgis/rest/services/nhd/MapServer):
+-   [nhd](https://hydro.nationalmap.gov/arcgis/rest/services/nhd/MapServer):
     A comprehensive set of digital spatial data that encodes information
     about naturally occurring and constructed bodies of surface water
     (lakes, ponds, and reservoirs), paths through which water flows
     (canals, ditches, streams, and rivers), and related entities such as
     point features (springs, wells, stream gauges, and dams).
-  - [govunits](https://carto.nationalmap.gov/arcgis/rest/services/govunits/MapServer):
+-   [govunits](https://carto.nationalmap.gov/arcgis/rest/services/govunits/MapServer):
     Major civil areas for the Nation, including States or Territories,
     counties (or equivalents), Federal and Native American areas,
     congressional districts, minor civil divisions, incorporated places
     (such as cities and towns), and unincorporated places.
-  - [contours](https://carto.nationalmap.gov/arcgis/rest/services/contours/MapServer):
+-   [contours](https://carto.nationalmap.gov/arcgis/rest/services/contours/MapServer):
     The USGS Elevation Contours service.
-  - [geonames](https://carto.nationalmap.gov/arcgis/rest/services/geonames/MapServer):
+-   [geonames](https://carto.nationalmap.gov/arcgis/rest/services/geonames/MapServer):
     Information about physical and cultural geographic features,
     geographic areas, and locational entities that are generally
     recognizable and locatable by name.
[…]
