### 2021-02-12 d587f35
Remove get_bbox (Fixes #14)
