+++ R/get_tiles.R
@@ -3,7 +3,7 @@
 #' This function splits the area contained within a bounding box into a set of
 #' tiles, and retrieves data from the USGS National map for each tile.
 #'
-#' @param bbox A bounding box representing the lower left and upper right corner
+#' @param data A bounding box representing the lower left and upper right corner
 #' of the area to retrieve a heightmap for. If not a
 #' \code{\link{terrainr_bounding_box}} object, it will be coerced to one.
 #' @param output_prefix The file prefix to use when saving tiles.
@@ -75,14 +75,15 @@
 #'   lng = runif(100, -74.01188, -73.83493)
 #' )
 #'
-#' bbox <- get_coord_bbox(lat = simulated_data$lat, lng = simulated_data$lng)
-#' bbox <- add_bbox_buffer(bbox, 100)
-#' get_tiles(bbox, tempfile())
+#' simulated_data <- sf::st_as_sf(simulated_data, coords = c("lng", "lat"))
+#'
+#' get_tiles(simulated_data, tempfile())
 #' }
 #'
 #' @export
 #' @md
-get_tiles <- function(bbox,
+#' @rdname get_tiles
+get_tiles <- function(data,
                       output_prefix = tempfile(),
                       side_length = NULL,
                       resolution = 1,
@@ -90,6 +91,125 @@ get_tiles <- function(bbox,
                       verbose = FALSE,
                       georeference = TRUE,
                       ...) {
+  UseMethod("get_tiles")
+}
+
+#' @rdname get_tiles
+#' @export
+get_tiles.sf <- function(data,
+                         output_prefix = tempfile(),
+                         side_length = NULL,
+                         resolution = 1,
+                         services = "elevation",
+                         verbose = FALSE,
+                         georeference = TRUE,
+                         ...) {
+
+  data <- sf::st_bbox(data)
+  bl <- c("lng" = data[["xmin"]], "lat" = data[["ymin"]])
+  tr <- c("lng" = data[["xmax"]], "lat" = data[["ymax"]])
+
+  get_tiles_internal(
+    bl = bl,
+    tr = tr,
+    output_prefix = output_prefix,
+    side_length = side_length,
+    resolution = resolution,
+    services = services,
+    verbose = verbose,
+    georeference = georeference,
+    ...
+  )
+
+}
+
+get_tiles.raster <- function(data,
+                             output_prefix = tempfile(),
+                             side_length = NULL,
+                             resolution = 1,
+                             services = "elevation",
+                             verbose = FALSE,
+                             georeference = TRUE,
+                             ...) {
+
+  data <- raster::extent(data)
+  bl <- c("lng" = data@xmin, "lat" = data@ymin)
+  tr <- c("lng" = data@xmax, "lat" = data@ymax)
+
+  get_tiles_internal(
+    bl = bl,
+    tr = tr,
+    output_prefix = output_prefix,
+    side_length = side_length,
+    resolution = resolution,
+    services = services,
+    verbose = verbose,
+    georeference = georeference,
+    ...
+  )
+
+}
+
+#' @rdname get_tiles
+#' @export
+get_tiles.list <- function(data,
+                           output_prefix = tempfile(),
+                           side_length = NULL,
+                           resolution = 1,
+                           services = "elevation",
+                           verbose = FALSE,
+                           georeference = TRUE,
+                           ...) {
+
+  bbox <- terrainr::terrainr_bounding_box(data[[1]], data[[2]])
+  get_tiles_internal(
+    bl = bbox@bl,
+    tr = bbox@tr,
+    output_prefix = output_prefix,
+    side_length = side_length,
+    resolution = resolution,
+    services = services,
+    verbose = verbose,
[…]
