### 2021-02-12 9a9c0ad
Shift centroid, add_bbox_buffer to sf (Fixes #18) (Fixes #17)
