+++ R/add_bbox_buffer.R
@@ -10,82 +10,183 @@
 #' [set_bbox_side_length] is a thin wrapper around [add_bbox_buffer] which sets
 #' all sides of the bounding box to (approximately) a specified length.
 #'
-#' @param bbox The original bounding box to add a buffer around. If not already
-#' a \code{\link{terrainr_bounding_box}} object, will be converted.
+#' @param data The original data to add a buffer around.
 #' @param distance The distance to add to the buffer.
 #' @param distance_unit The units of the distance to add to the buffer, passed
-#' to \code{\link{convert_distance}}.
-#' @param divisible Numeric: Extend the top right and bottom left corners so
-#' that each side is divisible by \code{divisible}. Leave set to \code{NULL} to
-#' not extend. This argument is in the same units as \code{distance_unit}.
+#' to [units::as_units].
+#' @param error_crs Logical: Should this function error if `data` has no CRS?
+#' If `TRUE`, function errors; if `FALSE`, function quietly assumes EPSG:4326.
+#' If `NULL`, the default, function assumes EPSG:4326 with a warning.
 #'
-#' @return A [terrainr_bounding_box] object.
+#' @return An `sfc` object (from [sf::st_as_sfc]).
 #'
 #' @family utilities
 NULL
 
 #' @rdname addbuff
 #' @examples
-#' add_bbox_buffer(
-#'   list(
-#'     c(lat = 44.04905, lng = -74.01188),
-#'     c(lat = 44.17609, lng = -73.83493)
-#'   ),
-#'   10
-#' )
+#'
+#' df <- data.frame(
+#'   lat = c(44.04905, 44.17609),
+#'   lng = c(-74.01188, -73.83493)
+#'   )
+#'
+#' df_sf <- sf::st_as_sf(df, coords = c("lng", "lat"))
+#' df_sf <- sf::st_set_crs(df_sf, 4326)
+#'
+#' add_bbox_buffer(df_sf, 10)
+#'
 #' @export
 #' @md
-add_bbox_buffer <- function(bbox,
+add_bbox_buffer <- function(data,
                             distance,
                             distance_unit = "meters",
-                            divisible = NULL) {
-  if (!methods::is(bbox, "terrainr_bounding_box")) {
-    bbox <- terrainr::terrainr_bounding_box(bbox[[1]], bbox[[2]])
+                            error_crs = NULL) {
+
+  UseMethod("add_bbox_buffer")
+
+}
+
+#' @rdname addbuff
+#' @export
+add_bbox_buffer.sf <- function(data,
+                               distance,
+                               distance_unit = "meters",
+                               error_crs = NULL) {
+
+  if (is.na(sf::st_crs(data)$input)) {
+    if (is.null(error_crs)) {
+      warning("No CRS associated with input data. Assuming EPSG:4326.\n")
+    } else if (error_crs) {
+      stop("No CRS associated with input data.")
+    }
+    data <- sf::st_set_crs(data, 4326)
   }
 
-  centroid <- terrainr::get_bbox_centroid(bbox)
-  corner_distance <- terrainr::calc_haversine_distance(
-    centroid,
-    bbox@bl
+  units(distance) <- units::as_units(distance_unit)
+
+  bbox <- sf::st_bbox(data)
+  bbox_sfc <- sf::st_as_sfc(bbox)
+  units(distance) <- distance_unit
+  bbox <- tryCatch({
+    # force an error before the warning if it'll be a problem
+    ignored <- units::as_units("degree")
+    ignored + distance
+    # If distance will error, we're already in the second method now.
[…]
