+++ tests/testthat/test-4-merge_rasters.R
@@ -28,13 +28,18 @@ test_that("merge_raster expected errors error expectedly", {
 
 test_that("merge_raster files are identical no matter the filename", {
   skip_on_cran()
-  first_tile <- add_bbox_buffer(list(
-    tr = c(lat = 44.050030001, lng = -74.01164),
-    bl = c(lat = 44.05003, lng = -74.011640001)
-  ), 10)
+  df <- data.frame(
+    lat = c(44.050030001, 44.05003),
+    lng = c(-74.01164, -74.011640001)
+  )
+  df_sf <- sf::st_as_sf(df, coords = c("lng", "lat"))
+  df_sf <- sf::st_set_crs(df_sf, 4326)
+  first_tile <- add_bbox_buffer(df_sf, 10)
+  ft_bbox <- sf::st_bbox(first_tile)
   second_tile <- terrainr_bounding_box(
-    bl = c(first_tile@bl@lat, first_tile@tr@lng),
-    tr = point_from_distance(first_tile@tr, 10, 90)
+    bl = c(lat = ft_bbox[["ymin"]], lng = ft_bbox[["xmin"]]),
+    tr = point_from_distance(c(lat = ft_bbox[["ymax"]],
+                               lng = ft_bbox[["xmax"]]), 10, 90)
   )
   # assign the output tile filenames...
   tmptif <- vector("list")
