### 2021-02-12 mikemahoney218/terrainr#20 @mikemahoney218
hit_api
> # `hit_national_map_api.R`
>
>     * I suggest using `httr::RETRY("GET", ...)`  rather then a tryCatch and counters, this will greatly simplify your logic. For example:
>
>
> ```
> #Set UP
> url = httr::parse_url("https://elevation.nationalmap.gov/arcgis/rest/services/3DEPElevation/ImageServer/exportImage")
>
> query_arg <- list(
>       bboxSR = 4326,
>       imageSR = 4326,
>       size = paste(8000, 8000, sep = ","),
>       format = "tiff",
>       pixelType = "F32",
>       noDataInterpretation = "esriNoDataMatchAny",
>       interpolation = "+RSP_BilinearInterpolation",
>       f = "json"
>     )
>
> ##
> first_corner <- bbox@bl
> second_corner <- bbox@tr
>
> bbox_arg <- list(bbox = paste(
>     min(first_corner@lng, second_corner@lng),
>     min(first_corner@lat, second_corner@lat),
>     max(second_corner@lng, first_corner@lng),
>     max(second_corner@lat, first_corner@lat),
>     sep = ","
>   ))
>
> ####======
> #^all the above could be replaces with paste(st_bbox(...), collapse = ",")
> ```
>
> You have a lot of logic wrapping your `httr` calls which is essentially:
>
> ```
> res <- httr::GET(url, query = c(bbox_arg, query_arg))
> content = httr::content(res, type = "application/json")
> ```
>
> An identical, but way to build in some of that error catching and retrying would be to change it to:
>
> ```
> res2 <- httr::RETRY("GET", url,
>                    query = c(bbox_arg, query_arg),
>                    times = 10, pause_cap = 10)
