### 2021-02-12 mikemahoney218/terrainr#20 @mikemahoney218
I am almost entirely certain that httr::RETRY never returns for one of these endpoints, for reasons I never got to the bottom of -- worth finding the commit instead of running into the same issue again
