### 2021-02-12 adcd83e
Shift classes to internal-only (Fixes #24)
