+++ R/raster_to_raw_tiles.R
@@ -17,6 +17,8 @@
 #'
 #' @examples
 #' \dontrun{
+#' if (!isTRUE(as.logical(Sys.getenv("CI")))) {
+#'
 #' simulated_data <- data.frame(
 #'   id = seq(1, 100, 1),
 #'   lat = runif(100, 44.04905, 44.17609),
@@ -27,6 +29,8 @@
 #' temptiff <- tempfile(fileext = ".tif")
 #' merge_rasters(output_files["elevation"][[1]], temptiff)
 #' raster_to_raw_tiles(temptiff, tempfile())
+#'
+#' }
 #' }
 #'
 #' @export
