+++ R/utils.R
@@ -32,44 +32,6 @@ rad_to_deg <- function(rad) {
   rad * 180 / base::pi
 }
 
-#' Convert distance into meters
-#'
-#' @param distance The numeric distance to be converted to meters
-#' @param distance_unit A string indicating the units to convert distance from
-#'
-#' @family utilities
-#'
-#' @return A numeric vector of distances, converted into meters.
-#'
-#' @examples
-#' convert_distance(100, "miles")
-#' @export
-convert_distance <- function(distance, distance_unit = c(
-                               "meters",
-                               "m",
-                               "kilometers",
-                               "km",
-                               "miles",
-                               "mi",
-                               "feet",
-                               "ft",
-                               "inches",
-                               "in"
-                             )) {
-  distance_unit <- distance_unit[[1]]
-  if (distance_unit == "miles" || distance_unit == "mi") {
-    distance * 1609.344
-  } else if (distance_unit == "feet" || distance_unit == "ft") {
-    distance * 0.3048
-  } else if (distance_unit == "kilometers" || distance_unit == "km") {
-    distance * 1000
-  } else if (distance_unit == "inches" || distance_unit == "in") {
-    distance * 0.0254
-  } else {
-    distance
-  }
-}
-
 #' Get the great-circle centroid for latitude/longitude data
 #'
 #' @param lat A vector of latitudes in degrees.
