+++ man/convert_distance.Rd
@@ -1,36 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/utils.R
-\name{convert_distance}
-\alias{convert_distance}
-\title{Convert distance into meters}
-\usage{
-convert_distance(
-  distance,
-  distance_unit = c("meters", "m", "kilometers", "km", "miles", "mi", "feet", "ft",
-    "inches", "in")
-)
-}
-\arguments{
-\item{distance}{The numeric distance to be converted to meters}
-
-\item{distance_unit}{A string indicating the units to convert distance from}
-}
-\value{
-A numeric vector of distances, converted into meters.
-}
-\description{
-Convert distance into meters
-}
-\examples{
-convert_distance(100, "miles")
-}
-\seealso{
-Other utilities: 
-\code{\link{addbuff}},
-\code{\link{calc_haversine_distance}()},
-\code{\link{deg_to_rad}()},
-\code{\link{get_centroid}()},
-\code{\link{point_from_distance}()},
-\code{\link{rad_to_deg}()}
-}
-\concept{utilities}
