### 2021-02-15 mikemahoney218/terrainr#20 @mikemahoney218
Found the commit at least, but past me was not kind enough to leave comments about the issue
https://github.com/mikemahoney218/terrainr/commit/f535f4bfa5bdbb4c2f6637617404b9291891f0d7
