### 2021-02-15 mikemahoney218/terrainr#20 @mikemahoney218
![Screenshot from 2021-02-15 10-56-24](https://user-images.githubusercontent.com/38229299/107968601-8d2b8700-6f7c-11eb-9402-4930600b3f53.png)

Ok, found (part of) why I removed httr::RETRY initially. There's a real performance impact when using RETRY for reasons I don't entirely understand -- there's a huge performance hit to RETRY vs my manual backoff implementation, and RETRY occasionally fails with very unfriendly (to me and the debugger) stacks (while it's downright rare to see similar failures in the current version)
