### 2021-02-15 mikemahoney218/terrainr#20 @mikemahoney218
(top is retry rewrite, bottom is mine)
