### 2021-02-15 mikemahoney218/terrainr#20 @mikemahoney218
Implemented httr::RETRY on branch:
https://github.com/mikemahoney218/terrainr/tree/httr_retry

CI has not been kind to this implementation (which I'm legitimately surprised by, since "it ran on my machine"). Notably, the tests are failing on different steps -- the code essentially "works", just is incredibly flaky:

* [NAIP server, code line 3](https://github.com/mikemahoney218/terrainr/runs/1904719012?check_suite_focus=true)
* [NAIP server, code line 23](https://github.com/mikemahoney218/terrainr/runs/1904719032?check_suite_focus=true)

Even ignoring the test suite failures, it's worth noting the time difference between the partial test suite runs with httr::RETRY versus without. I pushed a quick commit to reviewer_comments in order to get closer comparisons for the timing (adjust for server load etc) -- while it feels rude to benchmark someone else's public API, I'm fine with engineering a "natural experiment" :)

The timings below only reflect R CMD check, which isn't cached between builds.

* Partial test suite runs (httr::RETRY):
    * [4 min 5s](https://github.com/mikemahoney218/terrainr/runs/1904719012?check_suite_focus=true)
    * [4 min 27s](https://github.com/mikemahoney218/terrainr/runs/1904719032?check_suite_focus=true)
    * [4 min 11s](https://github.com/mikemahoney218/terrainr/runs/1904718952?check_suite_focus=true)
    * At the moment, the only suite to not fail is already [over 17 minutes](https://github.com/mikemahoney218/terrainr/runs/1904719005?check_suite_focus=true).
* Full test suite runs (manual backoff):
    * [1 min 53s](https://github.com/mikemahoney218/terrainr/runs/1904808463?check_suite_focus=true)
    * [2 min 14s](https://github.com/mikemahoney218/terrainr/runs/1904808396?check_suite_focus=true)
    * [2 min 9s](https://github.com/mikemahoney218/terrainr/runs/1904808464?check_suite_focus=true)
    * [3 min 5s](https://github.com/mikemahoney218/terrainr/runs/1904808403?check_suite_focus=true)

I run the same backoff protocol as httr does, so it _shouldn't_ be related to that.

I'm _pretty_ sure that part of the reason for both the additional failures and the increased run time is that get_href is solving for both res and body -- the two _combined_ get 15 tries, as if the first stage fails there's no point in advancing to the second. That said, you still see a significant time hit just by changing the last GET to RETRY (which really _should_ be an in-place substitution) so :shrug: I honestly don't know. I'm going to stick with the thing that works.
