+++ README.Rmd
@@ -30,14 +30,21 @@ At the absolute simplest level, terrainr provides a convenient and consistent AP
 ```{r eval=FALSE}
 library(terrainr)
 library(sf)
+
+# Optional way to display a progress bar while your tiles download:
+library(progressr)
+handlers("progress")
+
 simulated_data <- data.frame(id = seq(1, 100, 1),
                              lat = runif(100, 44.04905, 44.17609), 
                              lng = runif(100, -74.01188, -73.83493))
 
 simulated_data <- st_as_sf(simulated_data, coords = c("lng", "lat"))
 
-output_tiles <- get_tiles(simulated_data,
-                          services = c("elevation", "ortho"))
+with_progress( # Only needed if you're using progressr
+  output_tiles <- get_tiles(simulated_data,
+                            services = c("elevation", "ortho"))
+)
 ```
 
 ```{r eval=FALSE}
