### 2021-02-16 547984c
Wildly simplify merge_rasters (#22)
