+++ R/merge_rasters.R
@@ -2,21 +2,15 @@
 #'
 #' Some functions like [get_tiles] return multiple separate files
 #' when it can be useful to have a single larger raster instead. This function
-#' will collapse those multiple raster files into a single TIFF.
+#' is a thin wrapper over [sf::gdal_utils(util = "warp")], making it easy to
+#' collapse those multiple raster files into a single TIFF.
 #'
 #' @param input_rasters A character vector containing the file paths to the
 #' georeferenced rasters you want to use.
 #' @param output_raster The file path to save the merged georeferenced raster
 #' to. Must be a TIFF file. Ignored if \code{merge_raster} is not \code{TRUE}.
-#' @param input_images A character vector, the same length as
-#' \code{input_rasters}, containing the file paths to non-georeferenced images
-#' to merge. It's assumed that each image in the vector corresponds to the
-#' raster in the same position in input_rasters, and these images will be
-#' assigned the same extent and CRS. If you don't want to merge any images,
-#' leave as \code{NULL}.
-#' @param output_image The file path to save the merged images to.
-#' Must be a TIFF file. Leave as \code{NULL} if you aren't merging images.
-#' @param overwrite Logical: overwrite the output files if they exist?
+#' @param options Optionally, a character vector of options to be passed
+#' directly to [sf::gdal_utils].
 #'
 #' @return A named list containing the file paths outputs were written to.
 #'
@@ -24,14 +18,14 @@
 #'
 #' @examples
 #' \dontrun{
-#' tiles <- data.frame(
+#' simulated_data <- data.frame(
 #' lat = c(44.10379, 44.17573),
 #' lng = c(-74.01177, -73.91171)
 #' )
 #'
-#' tiles <- sf::st_as_sf(tiles, coords = c("lng", "lat"))
+#' simulated_data <- sf::st_as_sf(simulated_data, coords = c("lng", "lat"))
 #'
-#' img_files <- get_tiles(tiles)
+#' img_files <- get_tiles(simulated_data)
 #' merge_rasters(img_files[[1]])
 #'
 #' }
@@ -39,167 +33,12 @@
 #' @export
 merge_rasters <- function(input_rasters,
                           output_raster = tempfile(fileext = ".tif"),
-                          input_images = NULL,
-                          output_image = tempfile(fileext = ".tif"),
-                          overwrite = TRUE) {
-  output_list <- vector("list")
-  output_list$output_raster <- output_raster
+                          options = character(0)) {
 
-  # writeRaster seems to write a .tif if a .tiff is specified, which means
-  # mosaic_rasters does nothing and you get a useless blank .tif output unless
-  # you run the function twice.
-  # this is silly.
-  # so, we'll work around it if necessary -- do our work in a .tif then rename
-  # it at the end
-  if ((!grepl("\\.tif", output_raster)) ||
-    (!is.null(output_image) && !grepl("\\.tif?f$", output_image))) {
-    stop("Output files must be TIFFs.")
-  }
+  sf::gdal_utils(util = "warp",
+                 source = as.character(input_rasters),
+                 destination = output_raster,
+                 options = options)
 
-  fix_height <- 0
-  fix_ortho <- 0
-
-  if (grepl("\\.tiff$", output_raster)) {
-    fix_height <- 1
-    output_raster <- substr(output_raster, 1, nchar(output_raster) - 1)
-  }
-
-  input_raster_objects <- lapply(input_rasters, function(x) raster::raster(x))
-
-  # if some files were downloaded as RGBA and some RGB, mosaic_rasters will fail
-  #
[…]
