+++ NAMESPACE
@@ -5,7 +5,7 @@ S3method(add_bbox_buffer,sf)
 S3method(get_tiles,Raster)
 S3method(get_tiles,list)
 S3method(get_tiles,sf)
-S3method(get_tiles,terrainr_bounding_box)
+S3method(get_tiles,sfc)
 S3method(set_bbox_side_length,Raster)
 S3method(set_bbox_side_length,sf)
 export(add_bbox_buffer)

[…]
