### 2021-02-16 08c2dc1
Bump up coverage (Fixes #27)
