+++ tests/testthat/test-add_bbox_buffer.R
@@ -5,9 +5,26 @@ test_that("set_bbox_side_length works within 1%", {
   )
 
   df_sf <- sf::st_as_sf(df, coords = c("lng", "lat"))
+
+  expect_warning(
+    set_bbox_side_length(df_sf, 8000),
+    "No CRS associated with input data. Assuming EPSG:4326.\n"
+  )
+  expect_error(
+    set_bbox_side_length(df_sf, 8000, error_crs = TRUE),
+    "No CRS associated with input data."
+  )
+  bbox_default <- set_bbox_side_length(df_sf, 8000, error_crs = FALSE)
+
   df_sf <- sf::st_set_crs(df_sf, 4326)
 
   bbox <- set_bbox_side_length(df_sf, 8000)
+
+  expect_equal(
+    bbox_default,
+    bbox
+  )
+
   bbox <- sf::st_bbox(bbox)
 
   tl <- c(lat = bbox[["ymax"]], lng = bbox[["xmin"]])
@@ -25,4 +42,32 @@ test_that("set_bbox_side_length works within 1%", {
   8000,
   tolerance = 8000 * 0.005
   )
+
+  tmp_raster <- raster::raster("testdata/merge_rasters_test.tif")
+  rstr_bbox <- set_bbox_side_length(tmp_raster, 8000)
+  rstr_bbox <- sf::st_bbox(rstr_bbox)
+
+  tl <- c(lat = rstr_bbox[["ymax"]], lng = rstr_bbox[["xmin"]])
+  expect_equal(calc_haversine_distance(
+    c(lat = rstr_bbox[["ymax"]], lng = rstr_bbox[["xmax"]]),
+    tl
+  ),
+  8000,
+  tolerance = 8000 * 0.005
+  )
+  expect_equal(calc_haversine_distance(
+    c(lat = rstr_bbox[["ymin"]], lng = rstr_bbox[["xmin"]]),
+    tl
+  ),
+  8000,
+  tolerance = 8000 * 0.005
+  )
+
+  expect_equal(
+    rstr_bbox,
+    sf::st_bbox(add_bbox_buffer(tmp_raster,
+                                sqrt((8000^2) * 2) / 2)),
+    tolerance = 0.0001
+  )
+
 })
