+++ tests/testthat/test-get_tiles.R
@@ -77,3 +77,12 @@ test_that("split_bbox works consistently", {
     tolerance = 1
   )
 })
+
+test_that("raster method is consistent", {
+  tmp_raster <- raster::raster("testdata/merge_rasters_test.tif")
+  rstr_tile <- get_tiles(tmp_raster)
+  downloaded_raster <- raster::raster(rstr_tile[["elevation"]])
+  test_raster <- raster::raster("testdata/raster_tile.tif")
+  expect_equal(downloaded_raster@crs, test_raster@crs)
+  expect_equal(downloaded_raster@extent, test_raster@extent)
+})
