### 2021-02-17 e557c7f
Add geom_spatial_rgb (Fixes #28)
