+++ R/geom_spatial_rgb.R
@@ -1,0 +1,349 @@
+#' Plot RGB rasters in ggplot2
+#'
+#' `geom_spatial_rgb` and `stat_spatial_rgb` allow users to plot three-band RGB
+#' rasters in [ggplot2], using these layers as background base maps for other
+#' spatial plotting. Note that unlike [ggplot2::geom_sf], this function does
+#' _not_ force [ggplot2::coord_sf]; for accurate mapping, add
+#' [ggplot2::coord_sf] with a `crs` value matching your input raster as a layer.
+#'
+#' @rdname geom_spatial_rgb
+#' @param data The data to be displayed in this layer. In addition to the three
+#' options described in [ggplot2::geom_raster], there are two additional
+#' methods:
+#'
+#' If a `RasterStack` object (see [raster::stack]), this function will coerce
+#' the stack to a data frame and assume the raster bands are in RGB order
+#' (while allowing for, but ignoring, a fourth alpha band).
+#'
+#' If a length-1 character vector, this function will attempt to load the object
+#' via [raster::stack].
+#'
+#' @inheritParams ggplot2::geom_raster
+#' @param scale Integer. Maximum (possible) value in the three channels.
+#' If `NULL`, attempts to infer proper values from data -- if all RGB values
+#' are <= 1 then 1, <= 255 then 255, and otherwise 65535.
+#'
+#' @examples
+#' \dontrun{
+#'
+#' simulated_data <- data.frame(id = seq(1, 100, 1),
+#'                              lat = runif(100, 44.04905, 44.17609),
+#'                              lng = runif(100, -74.01188, -73.83493))
+#'
+#' simulated_data <- st_as_sf(simulated_data, coords = c("lng", "lat"))
+#'
+#' output_tiles <- get_tiles(simulated_data,
+#'                           services = c("ortho"),
+#'                           resolution = 120)
+#' )
+#'
+#' merged_ortho <- tempfile(fileext = ".tif")
+#' merge_rasters(output_tiles[["ortho"]], merged_ortho)
+#'
+#' merged_stack <- raster::stack(merged_ortho)
+#'
+#' ggplot() +
+#'   geom_spatial_rgb(data = merged_ortho),
+#'                    mapping = aes(x = x,
+#'                                  y = y,
+#'                                  r = red,
+#'                                  g = green,
+#'                                  b = blue)) +
+#'   geom_sf(data = simulated_data) +
+#'   coord_sf(crs = 4326)
+#'
+#' ggplot() +
+#'   geom_spatial_rgb(data = merged_stack),
+#'                    mapping = aes(x = x,
+#'                                  y = y,
+#'                                  r = red,
+#'                                  g = green,
+#'                                  b = blue)) +
+#'   geom_sf(data = simulated_data) +
+#'   coord_sf(crs = 4326)
+#'
+#' }
+#'
+#' @export
+geom_spatial_rgb <- function(mapping = NULL,
+                             data = NULL,
+                             stat = "spatialRGB",
+                             position = "identity",
+                             ...,
+                             hjust = 0.5,
+                             vjust = 0.5,
+                             interpolate = FALSE,
+                             na.rm = FALSE,
+                             show.legend = NA,
+                             inherit.aes = TRUE,
+                             scale = NULL) {
+  if (!(is.numeric(hjust) && length(hjust) == 1)) {
+    stop("`hjust` must be a numeric scalar")
+  }
+  if (!(is.numeric(vjust) && length(vjust) == 1)) {
+    stop("`vjust` must be a numeric scalar")
+  }
+  geom_spatial_rgb_internal(
+    data = data,
+    mapping = mapping,
+    stat = stat,
+    position = position,
+    ...,
+    hjust = hjust,
+    vjust = vjust,
+    interpolate = interpolate,
[…]
