+++ NEWS.md
@@ -25,9 +25,12 @@
     * A handful of utility functions are no longer exported:
         * `calc_haversine_distance`
         * `point_from_distance`
-* Improvements and bug fixes:
+* New features:
+    * Two new functions, `geom_spatial_rgb` and `stat_spatial_rgb`, allow you to 
+      use RGB map tiles as backgrounds for further plotting.
     * `calc_haversine_distance` gains an argument `coord_units` allowing it to 
       handle coordinates in radians as well as degrees.
+* Improvements and bug fixes:
     * `georeference_overlay` provides `tempfile(fileext = ".tif")` as a default
       output location if no `output_file` is provided.
     * `get_tiles` now tells you what tiles it's retrieving, not retriving.
@@ -49,6 +52,8 @@
     * `gdalUtilities` has been removed, with functionality replaced by `sf`.
     * `rlang` has been removed, with functionality removed.
     * `units` has been added.
+    * `ggplot2` has been moved to Imports (was previously in Suggests) due to 
+      the new `geom_spatial_rgb` and `stat_spatial_rgb` functions.
 
 # terrainr 0.2.1
 * Improvements and bug fixes:
