+++ R/geom_spatial_rgb.R
@@ -30,20 +30,22 @@
 #'                              lat = runif(100, 44.04905, 44.17609),
 #'                              lng = runif(100, -74.01188, -73.83493))
 #'
-#' simulated_data <- st_as_sf(simulated_data, coords = c("lng", "lat"))
+#' simulated_data <- sf::st_as_sf(simulated_data, coords = c("lng", "lat"))
+#' simulated_data <- sf::st_set_crs(simulated_data, 4326)
 #'
 #' output_tiles <- get_tiles(simulated_data,
 #'                           services = c("ortho"),
 #'                           resolution = 120)
-#' )
 #'
 #' merged_ortho <- tempfile(fileext = ".tif")
 #' merge_rasters(output_tiles[["ortho"]], merged_ortho)
 #'
 #' merged_stack <- raster::stack(merged_ortho)
 #'
+#' library(ggplot2)
+#'
 #' ggplot() +
-#'   geom_spatial_rgb(data = merged_ortho),
+#'   geom_spatial_rgb(data = merged_ortho,
 #'                    mapping = aes(x = x,
 #'                                  y = y,
 #'                                  r = red,
@@ -53,7 +55,7 @@
 #'   coord_sf(crs = 4326)
 #'
 #' ggplot() +
-#'   geom_spatial_rgb(data = merged_stack),
+#'   geom_spatial_rgb(data = merged_stack,
 #'                    mapping = aes(x = x,
 #'                                  y = y,
 #'                                  r = red,
