+++ man/terrainr_bounding_box.Rd
@@ -8,17 +8,14 @@ terrainr_bounding_box(bl, tr, coord_units = "degrees")
 }
 \arguments{
 \item{bl, tr}{The bottom left (\code{bl}) and top right (\code{tr}) corners of
-the bounding box, either as a \link{terrainr_coordinate_pair} object
+the bounding box, either as a [terrainr_coordinate_pair] object
 or a coordinate pair. If the coordinate pair is not named, it is assumed to
 be in (lat, lng) format; if it is named, the function will attempt to
 properly identify coordinates.}
 
-\item{coord_units}{Arguments passed to \link{terrainr_coordinate_pair}.
-If \code{bl} and \code{tr} are already \link{terrainr_coordinate_pair}
+\item{coord_units}{Arguments passed to [terrainr_coordinate_pair].
+If \code{bl} and \code{tr} are already [terrainr_coordinate_pair]
 objects, these arguments are not used.}
-}
-\value{
-A terrainr_bounding_box object
 }
 \description{
 In order to simplify code, most \code{terrainr} functions expect a set S4
@@ -26,3 +23,4 @@ class representation of coordinate pairs and bounding boxes. If the provided
 data is not in the expected S4 format, these functions are used to cast the
 data into the target class.
 }
+\keyword{internal}
