+++ man/get_centroid.Rd
@@ -22,7 +22,6 @@ Other utilities:
 \code{\link{addbuff}},
 \code{\link{calc_haversine_distance}()},
 \code{\link{deg_to_rad}()},
-\code{\link{point_from_distance}()},
 \code{\link{rad_to_deg}()}
 }
 \concept{utilities}
