### 2021-02-17 2037288
Validate documentation (Fixes #26)
