+++ R/utils.R
@@ -2,6 +2,8 @@
 #'
 #' @param deg A vector of values, in decimal degrees, to convert to radians
 #'
+#' @keywords internal
+#'
 #' @family utilities
 #'
 #' @return A vector of the same length in radians
@@ -9,7 +11,6 @@
 #' @examples
 #' deg_to_rad(360)
 #' rad_to_deg(deg_to_rad(360))
-#' @export
 deg_to_rad <- function(deg) {
   stopifnot(is.numeric(deg))
   deg * base::pi / 180
@@ -19,6 +20,8 @@ deg_to_rad <- function(deg) {
 #'
 #' @param rad A vector of values, in radians, to convert to decimal degrees
 #'
+#' @keywords internal
+#'
 #' @family utilities
 #'
 #' @return A vector of the same length in decimal degrees
@@ -26,7 +29,6 @@ deg_to_rad <- function(deg) {
 #' @examples
 #' rad_to_deg(2 * base::pi)
 #' rad_to_deg(deg_to_rad(360))
-#' @export
 rad_to_deg <- function(rad) {
   stopifnot(is.numeric(rad))
   rad * 180 / base::pi
