### 2021-02-17 3b3b67c
Test for get_tiles list names (Fixes #13)
