+++ tests/testthat/test-2-get_tiles_transportation.R
@@ -8,6 +8,8 @@ test_that("get_tiles gets the same transportation tiles twice", {
   georeference = FALSE
   )
 
+  expect_equal(names(output_tif), "transportation")
+
   expect_equal(length(output_tif), 1)
   expect_equal(length(output_tif[[1]]), 1)
 
