+++ tests/testthat/test-2-get_tiles_govunits.R
@@ -8,6 +8,8 @@ test_that("get_tiles gets the same govunits tiles twice", {
   georeference = FALSE
   )
 
+  expect_equal(names(output_tif), "govunits")
+
   expect_equal(length(output_tif), 1)
   expect_equal(length(output_tif[[1]]), 1)
 
