+++ tests/testthat/test-2-get_tiles_naip.R
@@ -8,6 +8,8 @@ test_that("get_tiles gets the same ortho tiles twice", {
   georeference = FALSE
   )
 
+  expect_equal(names(output_tif), "ortho")
+
   expect_equal(length(output_tif), 1)
   expect_equal(length(output_tif[[1]]), 1)
 
