+++ man/deg_to_rad.Rd
@@ -15,10 +15,6 @@ A vector of the same length in radians
 \description{
 Convert decimal degrees to radians
 }
-\examples{
-deg_to_rad(360)
-rad_to_deg(deg_to_rad(360))
-}
 \seealso{
 Other utilities: 
 \code{\link{addbuff}},
