+++ tests/testthat/test-calc_haversine_distance.R
@@ -19,4 +19,3 @@ test_that("Distance calculations make sense", {
     )
   )
 })
-
