### 2021-02-17 83f1839
Fix documentation bugs (Fixes #9) (Fixes #16) (Fixes #26)
