+++ tests/testthat/test-geom_spatial_rgb.R
@@ -15,17 +15,13 @@ test_that("all methods of geom_spatial_rgb are equivalent", {
   merge_rasters(output_tiles[["ortho"]], merged_ortho)
 
   test <- raster::stack(merged_ortho)
-  test_df <- as.data.frame(as(test, "SpatialPixelsDataFrame"))
+  test_df <- raster::as.data.frame(test, xy = TRUE)
+  test_df <- setNames(test_df, c("x", "y", "red", "green", "blue"))
 
   plots <- vapply(1:6, function(x) tempfile(fileext = ".png"), character(1))
 
   ggplot2::ggplot() +
-    geom_spatial_rgb(data = stats::setNames(test_df,
-                                            c("red",
-                                              "green",
-                                              "blue",
-                                              "x",
-                                              "y")),
+    geom_spatial_rgb(data = test_df,
                      mapping = ggplot2::aes(x = x,
                                             y = y,
                                             r = red,
@@ -55,12 +51,7 @@ test_that("all methods of geom_spatial_rgb are equivalent", {
     ggplot2::ggsave(plots[[3]])
 
   ggplot2::ggplot() +
-    stat_spatial_rgb(data = stats::setNames(test_df,
-                                            c("red",
-                                              "green",
-                                              "blue",
-                                              "x",
-                                              "y")),
+    stat_spatial_rgb(data = test_df,
                      mapping = ggplot2::aes(x = x,
                                             y = y,
                                             r = red,
