+++ vignettes/overview.Rmd
@@ -14,13 +14,13 @@ knitr::opts_chunk$set(
 )
 ```
 
-The goal of the terrainr package is to make it easy for users to download 
-geospatial data for points within the United States from the USGS National Map 
-API and then process that data for 
-further analysis, or to produce tiles that can then be imported into the 
-Unity 3D rendering engine to be used as physics objects. This vignette will 
-walk through the core functions available in terrainr and talk about a few 
-quirks of the National Map service and how to work around them.
+The goal of the terrainr package is to make it easier to visualize landscapes,
+both by providing functions to download elevation data and base maps from the 
+USGS National Map and by adding utilities to manipulate base maps for use in 
+visualizations using [ggplot2](https://ggplot2.tidyverse.org/) or the 
+freely available [Unity 3D rendering engine](https://unity.com/).
+This vignette will walk through the core functions available in terrainr and 
+how they interact.
 
 Let's load the terrainr package to get started:
 
@@ -32,7 +32,7 @@ We're going to work with data for Mount Elbert today, the highest point in the
 Rocky Mountain range. I'm just choosing this location for the dramatic scenery;
 the National Map can be used to retrieve data for the entire United States and 
 much of Canada. Let's simulate some data for the area right around Mt. Elbert,
-such as the point data we might get from field data:
+such as the point data we might get from some field collection:
 
 ```{r}
 mt_elbert_points <- data.frame(
@@ -41,31 +41,26 @@ mt_elbert_points <- data.frame(
 )
 ```
 
-Let's also assign the coordinates for the peak to `mt_elbert`:
-
-```{r}
-mt_elbert <- c(39.1178, -106.4452)
-```
-
-terrainr is built to play nicely with functions from the sf and raster packages.
-In order to get our simulated points into the right format, we need to use the 
-`st_as_sf` function from the `sf` package:
+terrainr is built to play nicely with functions from the 
+[sf](https://r-spatial.github.io/sf/) and [raster](https://rspatial.org/raster/) 
+packages. In order to get our simulated points into the right format, we need to 
+use the `st_as_sf` function from the `sf` package:
 
 ```{r}
 mt_elbert_points <- sf::st_as_sf(mt_elbert_points,
                                  coords = c("lng", "lat"))
+mt_elbert_points <- sf::st_set_crs(mt_elbert_points, 4326)
 ```
 
 Now that we've got our data in the right format, it's time to retrieve our data. 
 terrainr currently supports downloading DEMs from the USGS 3D Elevation Program 
 as well as orthoimages from the National Agricultural Imagery Program, in 
-addition to other raster data from the National Map. These programs each have
-slightly different APIs and different restrictions on file types and the size
-of image you can download at once. Rather than make you think about this, 
-terrainr handles all the weird edges of the API requests for you, including 
-splitting your request into tiles and downloading tiles separately. To make
-multi-step processing easier, terrainr functions which deal with these tiles 
-typically return lists of the file paths they saved your data to.
+addition to other base map images from the National Map. These programs each 
+have slightly different APIs and different restrictions on file types and the 
+size of image you can download at once. 
+Rather than make you think about this, terrainr handles all the edges of making 
+API requests for you, including splitting your request into tiles and 
+formatting the query. 
 
 For this vignette, we'll retrieve both elevation and orthoimagery using the 
[…]
