+++ NAMESPACE
@@ -20,4 +20,3 @@ export(raster_to_raw_tiles)
 export(set_bbox_side_length)
 export(stat_spatial_rgb)
 export(vector_to_overlay)
-importClassesFrom(sp,SpatialPixelsDataFrame)

[…]
