+++ README.md
@@ -135,46 +135,46 @@ you’ll be more confident that your computer is still churning along and
 not just stalled out. For more information, check out [the introductory
 vignette](https://mikemahoney218.github.io/terrainr/articles/overview.html)
 and [the guide to importing your data into
-Unity!](https://mikemahoney218.github.io/terrainr/articles/unity_instructions.html)
+Unity\!](https://mikemahoney218.github.io/terrainr/articles/unity_instructions.html)
 
 ## Available Datasets
 
 The following datasets can currently be downloaded using `get_tiles` or
 `hit_national_map_api`:
 
--   [3DEPElevation](https://elevation.nationalmap.gov/arcgis/rest/services/3DEPElevation/ImageServer):
+  - [3DEPElevation](https://elevation.nationalmap.gov/arcgis/rest/services/3DEPElevation/ImageServer):
     The USGS 3D Elevation Program (3DEP) Bare Earth DEM.
--   [USGSNAIPPlus](https://services.nationalmap.gov/arcgis/rest/services/USGSNAIPPlus/MapServer):
+  - [USGSNAIPPlus](https://services.nationalmap.gov/arcgis/rest/services/USGSNAIPPlus/MapServer):
     National Agriculture Imagery Program (NAIP) and high resolution
     orthoimagery (HRO).
--   [nhd](https://hydro.nationalmap.gov/arcgis/rest/services/nhd/MapServer):
+  - [nhd](https://hydro.nationalmap.gov/arcgis/rest/services/nhd/MapServer):
     A comprehensive set of digital spatial data that encodes information
     about naturally occurring and constructed bodies of surface water
     (lakes, ponds, and reservoirs), paths through which water flows
     (canals, ditches, streams, and rivers), and related entities such as
     point features (springs, wells, stream gauges, and dams).
--   [govunits](https://carto.nationalmap.gov/arcgis/rest/services/govunits/MapServer):
+  - [govunits](https://carto.nationalmap.gov/arcgis/rest/services/govunits/MapServer):
     Major civil areas for the Nation, including States or Territories,
     counties (or equivalents), Federal and Native American areas,
     congressional districts, minor civil divisions, incorporated places
     (such as cities and towns), and unincorporated places.
--   [contours](https://carto.nationalmap.gov/arcgis/rest/services/contours/MapServer):
+  - [contours](https://carto.nationalmap.gov/arcgis/rest/services/contours/MapServer):
     The USGS Elevation Contours service.
--   [geonames](https://carto.nationalmap.gov/arcgis/rest/services/geonames/MapServer):
+  - [geonames](https://carto.nationalmap.gov/arcgis/rest/services/geonames/MapServer):
     Information about physical and cultural geographic features,
     geographic areas, and locational entities that are generally
     recognizable and locatable by name.
--   [NHDPlus\_HR](https://hydro.nationalmap.gov/arcgis/rest/services/NHDPlus_HR/MapServer):
+  - [NHDPlus\_HR](https://hydro.nationalmap.gov/arcgis/rest/services/NHDPlus_HR/MapServer):
     A comprehensive set of digital spatial data comprising a nationally
     seamless network of stream reaches, elevation-based catchment areas,
     flow surfaces, and value-added attributes.
--   [structures](https://carto.nationalmap.gov/arcgis/rest/services/structures/MapServer):
+  - [structures](https://carto.nationalmap.gov/arcgis/rest/services/structures/MapServer):
     The name, function, location, and other core information and
     characteristics of selected manmade facilities.
--   [transportation](https://carto.nationalmap.gov/arcgis/rest/services/transportation/MapServer):
[…]
