### 2021-02-18 #416 @mikemahoney218
Thank you Mike for the thorough review! Below I've attempted to go through your comments line-by-line; I've collapsed comments that were repeated into single items (mostly httr/sf related) and removed the complements (though they were much appreciated :) ) for length. All the changes below are now merged to main as version 0.3.0.

> ## Key Points:
>
>     * I think there is some nice utility here but there is also some cross over with others packages like `FedData`, `elevatr`.
>
>     * Following this, other then elevation, this is not a "spatial data" retrieval package as it returns PNGs that are more similar to map tiles then to spatial data. Thus the functionality is also similar to `rosm`.
>
>     * I cannot think of many cases for wanting the PNG files in R. That **certainly** does not mean they are not out there but I think more clarity about what the package is returning (maptiles) and there use would help it gain traction.

I fully agree that the API code overlaps slightly with a few other packages (particularly the FedData package, which uses two of the same services). As far as the API wrapper code is concerned, I see the primary differences as being:

* 7 endpoints not serviced by any other R package.
* USGS data is released to the public domain, unlike OSM (which uses a strong copyleft license with attribution requirements that I personally would not be willing to use in any project; [link](https://www.openstreetmap.org/copyright)) and MapZen (which uses CC-BY-4.0 which at least removes the share-alike requirements; [link](https://github.com/tilezen/joerd/blob/master/docs/LICENSE.md)).

With that said, I think the main differentiator of terrainr is the workflow it introduces for creating 3D terrain tiles with overlays in the Unity engine. There's a lot of buzz in visualization research about using game engines for interactive simulations of landscape dynamics, which is why I originally started working on terrainr; the `get_tiles %>% merge_raster %>% raster_to_raw_tiles` flow reduces what used to be an intensely manual process into an all-but-automated method (see for instance this [five minute tutorial](https://www.youtube.com/watch?v=KiRq3SXhOc4) that walks through manually downloading a heightmap, wrangling the data in GIMP, and then importing it -- without any guidance on adding overlays).

I've clarified that most endpoints are returned as tiles. I've also attempted to rework documentation to emphasize that the data manipulation and export functionality are at least co-equal goals to the retrieval functions; unfortunately, as these functions need data to manipulate, the data retrieval functions have to before the manipulation functions in any given document. With that said, there are still a number of use cases for these image tiles within R -- see for instance [this example with rayshader](https://twitter.com/researchremora/status/1341220887836692480?s=20). I imagine the new `geom_spatial_rgb` (discussed below) will also be helpful here.

>     * Use `httr::RETRY` to handle much of your error catching and `httr::write_disk()` to avoid having to read images in as raw bytes/bits.  This will save a large amount of code as well.

I've attempted to use `httr::RETRY` in the place of my manual backoff before, with no success (I removed my last attempt in [October](https://github.com/mikemahoney218/terrainr/commit/f535f4bfa5bdbb4c2f6637617404b9291891f0d7)). I attempted again following this review ([documented here](https://github.com/mikemahoney218/terrainr/issues/20#issuecomment-779333950)), but some challenges with this API (sometimes returning failures with a 200 status, returning HTML in lieu of JSON, sometimes returning successes with a 400, and I'm sure other weird aberrant behaviors) can't be handled as well in httr::RETRY as in the manual code. The end result is various CI failures, and retrieving each tile takes notably longer than the current manual backoff approach. As a result, unfortunately I don't see a way to incorporate `httr::RETRY` at this time.

Following last month's [challenge with the transportation endpoint starting to return raw PNGs instead of base64 encoded images], I've changed the way `get_tiles` works so that instead of having a list of base64 vs raw image endpoints, `get_tiles` will check the type of the returned data instead ([first relevant commit](https://github.com/mikemahoney218/terrainr/commit/5c5d907984d4130cd17e147790ee338dc712efc2#diff-b1d5001a38d627f4fac6b61cf56218c2ec49377a47abc3436ddb6667f160b2e6L194-L224), [current version](https://github.com/mikemahoney218/terrainr/blob/851bcba8819e28084548edb84835bc71ed74bb03/R/get_tiles.R#L347-L356)). As a result, I believe I need to load the tiles into memory at some point in order to detect if they need to be decoded. Because of this, I don't think `httr::write_disk` will save system resources at the moment.

>     * With all of these I think you can drop: `base64enc`,  `gdalUtils`, `methods`, `gdalUtilities`, and `rlang` from your dependencies and many of your function can be slimmed down and/or removed. The result (in my option) will be a more lightweight package with higher utility, and more easy integration in spatial workflows.

At the end of the day, I was able to drop `methods`, `gdalUtilities`, and `rlang`; `gdalUtils` is still in use in `raster_to_raw_tiles` for reasons discussed b and `base64enc` is required for decoding endpoints that don't return data as a raw vector.

> ## Branding
>
>     * It's a weird point but one that I think is important. `terrainr` as a name sells your package short while also being misleading. Terrain is one of 9 endpoints offered but is also anomalous in that it is the only "spatial" data available in your package (in the traditional vector/raster sense).
>
>     * Arguably terrain is also the best served of the resources in the R ecosystem with `elevatr` and `FedData`. `Elevatr` on the surface is a better choice for getting elevation data simply because it offers multiple resolutions and the Mapzen server is quicker then the ArcGIS server.
>
>     * My expectation of the package coming in given the description of "retrieve geospatial data" was that for resources like the NHD, NHDHighRes, WBD, and transportation, I would be getting the vector data. Particularly since these are offered by the National Map. Therefore, it is very important to be upfront that the package is offering endpoints to (effectively) mosaicked web tiles (with one spatial data source).
>
>     * All of this is not to discount the utility of your package because I think it is one of the few to offer access to these resources which makes it very unique! But the functionality needs to be explicit because, as is, someone coming in with the expectation that they will be able to query spatial data will be disappointed, while those looking for map tile data will likely pass your package by.

This comment actually took me by surprise originally! It had never even occurred to me that someone would see the name `terrainr` and associate it with the elevation endpoints (though of course it makes sense now). The name harkens back to the original goal of the package: creating terrain tiles in Unity. As discussed in the [Unity vignette](https://mikemahoney218.github.io/terrainr/articles/unity_instructions.html), "terrain" is used as a very precise name for what's being created (physically-explicit heightmap tiles created from planar-interlaced raw image files) which is used throughout the Unity UI. As mentioned in the earlier comments, I've attempted to make this goal of the package more explicit in the documentation and DESCRIPTION file, as well as more explicitly expressing that data is obtained as tiles.

As a quick side note, I want to point out that `get_tiles` has an argument `resolution` which lets you set pixel side length in meters.

>     * One **huge** opportunity for your package, that I think you already have worked out, is how to integrated these resources with ggplot as a base-map. It is a constant frustration of mine (and many I know) to get nice base maps for ggplot.
>
>     * Yes, `tmap` has some utilities for this and `ggmap` is out there (I have always struggled with it), there is room here for growth. If you could add a `geom_terrainr()` that played nicely with `ggplot`, I think that would be immediately useful to a many people.

Added `geom_spatial_rgb` ([relevant commit](https://github.com/mikemahoney218/terrainr/commit/e557c7fc704bc10336a7398046645e56c780c9fe)) for displaying RGB tiles.

> # Code Review:
>
> ## `add_bbox_buffer.R`

I've adopted your suggestion about how to shift these to sf. The older behavior is preserved in order to manipulate latitude/longitude data (but still uses units and sf as suggested). {Relevant commit](https://github.com/mikemahoney218/terrainr/commit/9a9c0ad05ee7b67206f53823ad238d1d35eb85e9#diff-1a631b8a8a3538dc5aaf4db9172dddd880a76a05c92e33324d83e4d21db50eda).

> ## `classes.R`
> # `get_bbox.R`
> # `get_bbox_centroid`

These have been removed from all user-facing code. The classes still exist for communication between package functions, but are no longer the only accepted inputs for functions and are not returned by any function. `sf` or `Raster` objects can be used in their place across the board. [Relevant commit](https://github.com/mikemahoney218/terrainr/commit/3ba328b56bda664906b7eb740914d2cb16ef9f73).

> # `get_tiles.R`
>
>     * This function, even for the example, takes a while to run, it would be nice to have some messaging along the way to give confidence it is going if `progressr` is not installed. Consider using `httr::progress()` in `hit_national_map_api`. More on that below.

I personally am very persuaded by [progressr's motto](https://cran.r-project.org/web/packages/progressr/vignettes/progressr-intro.html):

> The developer is responsible for providing progress updates but it's only the end user who decides if, when, and how progress should be presented. No exceptions will be allowed.

Letting users control how and why they're updated is the key goal of using progressr, so I'm reluctant to add always-on progress bars (which would trigger three separate times per tile downloaded, I believe) via httr at the moment.

>     * Its nice to be able to provide a `tmpfile` prefix, but would be equally nice to be able to set where the file is downloaded to (other then the tmp directory). This will come up again below with the GDAL discussion.

Try setting `output_prefix` to any valid path on your system for this. `get_tiles` will append a string in the format `servicename_xtile_ytile.tif` to `output_prefix` for each tile downloaded.

>     * find a way to avoid the `3` in the beginning of your list name for 3DEP, maybe elevation3DEP?

I changed the behavior so that now `get_tiles` returns a list whose names match the names provided to `service`; this means that if you provide `services = "elevation"` the output list will now be named `elevation`. If a user specifies `3DEPElevation`, I think it still makes sense to name that element of the list `3DEPElevation` in order to have consistent behavior across endpoints; I do wish the official service name was something else! [Relevant commit](https://github.com/mikemahoney218/terrainr/commit/a800751ded03bc8ab29bca900cdd84887f8fd2de#diff-b1d5001a38d627f4fac6b61cf56218c2ec49377a47abc3436ddb6667f160b2e6).

> # `merge_rasters.R`

I've adopted your suggested sf/gdal_warp code entirely, and cannot _believe_ how much faster and cleaner it runs now. Thanks in particular for this suggestion! [Relevant commit](https://github.com/mikemahoney218/terrainr/commit/547984ce8d4ed3fee9e6e12bf52a3919b2240026).

> # `raster_to_raw_tiles.R`
>
> Use `sf::gdal_utils("translate", ...)` here instead of gdalUtils/gdalUtilities for the same reasons as above.

There's a comment in this file from about five months ago that I left as a warning to my future self:

https://github.com/mikemahoney218/terrainr/blob/851bcba8819e28084548edb84835bc71ed74bb03/R/raster_to_raw_tiles.R#L78

Unfortunately, this is still true. `gdalUtilities` is a simple wrapper around `sf::gdal_utils` (with command formatting more similar to command-line gdal), and so using `sf::gdal_utils` through the wrapper still causes my R session to hard crash for reasons I'm entirely unaware of. Unfortunately, that has left me dragging in `gdalUtils` for the time being.

> It may be worth your time to look at this: https://github.com/ropensci/tiler

Unfortunately, I'm yet to get this package to work on my laptop due to an issue with Python package dependencies.

> ### georeference
>
> The given example does not run because the arguments are in the wrong order, also it doesn't seem to need alignment? The input and output are identical.

Serves me right for not naming my arguments! [Addressed both issues here](https://github.com/mikemahoney218/terrainr/commit/741fb3544c366deee168034ca9747beda80e093c).

> No need for TIFF package (I think) since `raster` reads tiffs already (raster will also read PNG and JPEG, just into RGBA bands)

I had to do a little bit of investigation into why this function works this way, myself! It turns out that while `raster::brick` is perfectly happy to read in any type of file, it only correctly detects the scale of the bands if the file is entirely read into memory (and assumes 255/8-bit images if not). Given that `georeference_overlay` has to handle images with unknown scales (for instance, one user of this package edits the tiles in Photoshop) and even the default case (tiles returned by the National Map and `vector_to_overlay`) is scaled [0, 1], this function has to load the images into memory before creating RasterBricks to avoid all outputs being either black or white squares. I haven't fully investigated if this is true with TIFF images as well (verified for JPEG and PNG), but I believe the current approach is the simplest way to handle varied file types.

> # `utils.R`
>
> Check out the `units` package to handle all these unit conversions, but really, the unit given should match the CRS of the input data. Because of this, I might consider being more strict about the units you allow for a buffering call.

I've swapped over to `units` for this across the board (except in the case of radians -> degree conversions, where I still use my helper functions -- which are no longer exported -- to make the code a bit shorter/more legible). As someone working in a field (forestry) that still relies upon inches and chains as standard units, I'm reluctant to force people to SI standards.

> Notes:
>
>     * Readme - you have an 'elev' hanging at the end of the 2nd 'plot' line

Doh! [Fixed](https://github.com/mikemahoney218/terrainr/commit/0728a3280b340c6140c30da40ff9faad0f399ec8), thank you.
