### 2021-02-18 #416 @sfoks
---
### Provided Review Template
  *No relationship other than being asked to review this R package.*

Installation was okay after updating my Mac to new R, new devtools, etc. I didn't have access to my windows machine this weekend, otherwise I'm sure the installation would have been easier!

---
### Additional reviewer questions from guide
- Does the code comply with general principles in the Mozilla reviewing guide?
*I believe so*
- Does the package comply with the rOpenSci packaging guide?
*yes*
- Are there improvements that could be made to the code style?
*code style is fine*
- Is there code duplication in the package that should be reduced?
*not that I can comment to*
- Are there user interface improvements that could be made?
*NA*
- Are there performance improvements that could be made?
*not that I can comment to*
- Is the documentation (installation instructions/vignettes/examples/demos) clear and sufficient? Does it use the principle of multiple points of entry i.e. takes into account the fact that any piece of documentation may be the first encounter the user has with the package and/or the tool/data it wraps?
*I believe the documentation is more than sufficient. I was able to navigate the vignettes in my first encounter, had no errors when running examples, and only problems with navigating the software Unity.*
- Were functions and arguments named to work together to form a common, logical programming API that is easy to read, and autocomplete?
*yes*

---
- This review is for terrainr up to commit 91a73a4.
- I was using R version 4.0.3, Rstudio version 1.4.1103, and a macOS Big Sur 11.2.1 when I performed this review.
- Within the unity_instructions.Rmd, I think  `merged_outputs <- merge_rasters(raw_tiles$`3DEPElevation`, tempfile(fileext = ".tif"), .$USGSNAIPPlus, tempfile(fileext = ".tif"))` should be changed to: `merged_outputs <- merge_rasters(raw_tiles$`3DEPElevation`, tempfile(fileext = ".tif"), raw_tiles$USGSNAIPPlus, tempfile(fileext = ".tif"))`
- In general, I received a lot of warnings with the unity_instructions vignette, several related to the `merge_rasters` function. ex: ` Warning 6: gdalbuildvrt does not support heterogeneous band numbers: expected 3, got 4. Skipping /var/folders/zd/0gcdnc8x21177wtw966dqktm0000gn/T//Rtmp1XicTF/file2176d7036b2.tif `; however this didn't seem to influence the generation of output.
- Other warnings were received upon performing `raster_to_raw_tiles` function: ex `Warning 1: -srcwin 0 8194 4097 4097 falls partially outside raster extent. Going on however. ` I believe you explain that this is due to National Map API issues in trying to download a certain geographic extent but wanted to make sure you were aware of it. I also believe these warnings (also associated with the `raster_to_raw_tiles` function) are related to the same National Map API download issues or perhaps my instance of GDAL?
`Warning messages: 1: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 1, nodata value has been clamped to 0, the original value being out of range. 2: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 2, nodata value has been clamped to 0, the original value being out of range. 3: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 3, nodata value has been clamped to 0, the original value being out of range. 4: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 4, nodata value has been clamped to 0, the original value being out of range.`
- Give these warning examples listed above, the raw, xml, and png files exported to my directory correctly.
- I downloaded Unity to test terrainr output, but could not get my tiles into the software! This is an issue with me never have used Unity, so definitely a user-error. I may try this again at a later date since it seems like such a cool way to visualize this data.
- I think it would be great to present additional examples of data pulls and use in Unity, but of course this is not necessary for package publication. It appears there is a lot that this package can do and providing more examples of functionality to novice users like me would be awesome.
- Overall I think this package is well put together, thanks for allowing me to review!
