### 2021-02-18 #416 @mikemahoney218
Thank you so much @sfoks ! I've responded to your comments below.

>     * Within the unity_instructions.Rmd, I think  `merged_outputs <- merge_rasters(raw_tiles$`3DEPElevation`, tempfile(fileext = ".tif"), .$USGSNAIPPlus, tempfile(fileext = ".tif"))` should be changed to: `merged_outputs <- merge_rasters(raw_tiles$`3DEPElevation`, tempfile(fileext = ".tif"), raw_tiles$USGSNAIPPlus, tempfile(fileext = ".tif"))`

Fixed, thank you! [Relevant commit](https://github.com/mikemahoney218/terrainr/commit/83f183964aab7db782ead3b1dabb07a76409e191#diff-d4ca28daf787b164d1a2c9860d979482b7a151431a00bc49a48f45a3528d63e8R29-R42).

>     * In general, I received a lot of warnings with the unity_instructions vignette, several related to the `merge_rasters` function. ex: `Warning 6: gdalbuildvrt does not support heterogeneous band numbers: expected 3, got 4. Skipping /var/folders/zd/0gcdnc8x21177wtw966dqktm0000gn/T//Rtmp1XicTF/file2176d7036b2.tif`; however this didn't seem to influence the generation of output.

This _should_ be resolved thanks to changes made for Mike's review ([relevant commit](https://github.com/mikemahoney218/terrainr/commit/547984ce8d4ed3fee9e6e12bf52a3919b2240026#diff-1fcd7b95e21a5b535edfb22473c3092a800748fbe470feb73e2afaf1af90cc62R38-R41)). I've validated that this "works on my machine" without warnings thanks to the new and improved `merge_rasters` and haven't seen any warnings in CI output yet.

This used to happen when the National Map returned some image tiles with alpha bands and some without, due to a limitation in the workflow we were using; however, whether or not a given tile will have an alpha band _seems_ to change over time (so that running the same query a day later might resolve the issue). I _think_ this resulted in a tile not getting merged into your full raster, resulting in the `CPL_gdaltranslate` error mentioned below; if I'm right, then I think the commit above solves both issues!

>     * Other warnings were received upon performing `raster_to_raw_tiles` function: ex `Warning 1: -srcwin 0 8194 4097 4097 falls partially outside raster extent. Going on however. ` I believe you explain that this is due to National Map API issues in trying to download a certain geographic extent but wanted to make sure you were aware of it. I also believe these warnings (also associated with the `raster_to_raw_tiles` function) are related to the same National Map API download issues or perhaps my instance of GDAL?
>       `Warning messages: 1: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 1, nodata value has been clamped to 0, the original value being out of range. 2: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 2, nodata value has been clamped to 0, the original value being out of range. 3: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 3, nodata value has been clamped to 0, the original value being out of range. 4: In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 4, nodata value has been clamped to 0, the original value being out of range.`
>
>     * Give these warning examples listed above, the raw, xml, and png files exported to my directory correctly.

Thank you for flagging!

*Warning 1: -srcwin 0 8194 4097 4097 falls partially outside raster extent. Going on however*:

You're exactly right, this warning is related to the National Map endpoint not exactly respecting the requested bounding boxes. I've made this explicit in the vignette ([relevant commit](https://github.com/mikemahoney218/terrainr/commit/128e34e807eb1eb6943de8a69a485f39210b3a20)), but to explain here as well:

Since Unity expects to import heightmaps as squares with side lengths that are powers of 2, this function will create tiles that are the right size for Unity; in order to make things easier for the user, those tiles will all be the same size (so users don't need to change sizes for each tile they import). That means that if the input raster isn't exactly divisible by side_length, this function will create tiles with no data past a certain point ("falls partially outside raster extent").

I've left this message as standard output (rather than an official R warning) because I believe it's useful for users to know to expect some no-data regions at the edges of their terrain tiles, but it's standard function behavior (so shouldn't interrupt users who have set `options(warn=2)` or similar). I'm happy to change this behavior (either provide a message from within terrainr, rather than messages from GDAL, or raise an actual warning) if you think something else makes more sense!

*In CPL_gdaltranslate(source, destination, options, oo,  ... : GDAL Message 1: for band 1, nodata value has been clamped to 0, the original value being out of range.`:

I believe this is about the issue with mixed 3 & 4 band rasters as mentioned above, and should _hopefully_ be fixed given the new `merge_rasters` code.

>     * I downloaded Unity to test terrainr output, but could not get my tiles into the software! This is an issue with me never have used Unity, so definitely a user-error. I may try this again at a later date since it seems like such a cool way to visualize this data.

If you're willing, I'd love to hear what went wrong! I would love to make the unity_instructions vignette as comprehensive as possible for future users.

>     * I think it would be great to present additional examples of data pulls and use in Unity, but of course this is not necessary for package publication. It appears there is a lot that this package can do and providing more examples of functionality to novice users like me would be awesome.

Thank you! I fully hope and intend to keep adding examples to the documentation moving forward.

>     * Overall I think this package is well put together, thanks for allowing me to review!

Thank you so much for reviewing!
