### 2021-02-19 #416 @ldecicco-USGS
@mikemahoney218 thanks for detailed response!

@sfoks and @mikejohnson51  Could you please indicate whether you are happy with the authors' response? Template https://devguide.ropensci.org/approval2template.html
