### 2021-02-19 #416 @sfoks
## Reviewer Response

#### Final approval (post-review)

My earlier comment: `I downloaded Unity to test terrainr output, but could not get my tiles into the software! This is an issue with me never have used Unity, so definitely a user-error. I may try this again at a later date since it seems like such a cool way to visualize this data.`
Author response: `If you're willing, I'd love to hear what went wrong! I would love to make the unity_instructions vignette as comprehensive as possible for future users.`
- I was lost at the import_raw step. I followed your instructions, matched the screenshot values exactly, and tried the "Windows" option and "Mac" option (since I was on a Mac when testing terrainr). I loaded the raw file in both cases apparently, but neither seemed produce anything that looked like Mt. Marcy (the terrain tile was still highlighted in orange and it still had checker-boarded fill). So I'm not sure if I am missing a crucial step here. A screen shot of what a successful, first terrain tile looks like might help users or me perhaps, but really these are issues with using Unity and ultimately have no influence on publishing. I could probably cruise YouTube videos to see what I am doing wrong here. Basically no response is needed here in terms of adjusting the vignette, just adding more details to the problems I had with Unity.
- thanks for allowing me to review!

Estimated hours spent reviewing: 11 hours total
