+++ DESCRIPTION
@@ -13,8 +13,8 @@ Description: Functions for the retrieval, manipulation, and visualization of
     in the Unity '3D' rendering engine. Functions are also provided for 
     retrieving elevation data and base map tiles from the 'USGS' National Map
     ('<https://viewer.nationalmap.gov/services/>').
-URL: https://mikemahoney218.github.io/terrainr/, https://github.com/mikemahoney218/terrainr
-BugReports: https://github.com/mikemahoney218/terrainr/issues
+URL: hhttps://docs.ropensci.org/terrainr, https://github.com/ropensci/terrainr
+BugReports: https://github.com/ropensci/terrainr/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
