+++ DESCRIPTION
@@ -13,7 +13,7 @@ Description: Functions for the retrieval, manipulation, and visualization of
     in the Unity '3D' rendering engine. Functions are also provided for 
     retrieving elevation data and base map tiles from the 'USGS' National Map
     ('<https://viewer.nationalmap.gov/services/>').
-URL: hhttps://docs.ropensci.org/terrainr, https://github.com/ropensci/terrainr
+URL: https://docs.ropensci.org/terrainr, https://github.com/ropensci/terrainr
 BugReports: https://github.com/ropensci/terrainr/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
