### 2021-02-19 7063155
Fix a few URLs and image sizes
