+++ man/hit_national_map_api.Rd
@@ -40,7 +40,7 @@ This function retrieves a single tile of data from a single National Map
 service and returns the raw response. End users are recommended to use
 \link{get_tiles} instead, as it does much more validation and provides
 a more friendly interface. For a description of the datasets provided by the
-National Map, see \url{https://viewer.nationalmap.gov/services/}
+National Map, see \url{https://apps.nationalmap.gov/services}
 }
 \section{Additional Arguments}{
 
