+++ DESCRIPTION
@@ -20,8 +20,8 @@ Description: Functions for the retrieval, manipulation, and visualization of
     geospatial data, with an aim towards producing '3D' landscape visualizations
     in the Unity '3D' rendering engine. Functions are also provided for 
     retrieving elevation data and base map tiles from the 'USGS' National Map
-    ('<https://viewer.nationalmap.gov/services/>').
-URL: https://docs.ropensci.org/terrainr, https://github.com/ropensci/terrainr
+    ('<https://apps.nationalmap.gov/services/>').
+URL: https://docs.ropensci.org/terrainr/, https://github.com/ropensci/terrainr
 BugReports: https://github.com/ropensci/terrainr/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
