+++ _pkgdown.yml
@@ -1,4 +1,4 @@
-url: https://docs.ropensci.org/terrainr
+url: https://docs.ropensci.org/terrainr/
 authors:
   Mike Mahoney:
     href: https://www.mm218.dev/
