### 2021-02-19 be8c677
Few final spelling changes pre-CRAN
