+++ man/terrainr-package.Rd
@@ -9,10 +9,10 @@
 \if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
 
 Functions for the retrieval, manipulation, and visualization of 
-    geospatial data, with an aim towards producing '3D' landscape visualizations
-    in the Unity '3D' rendering engine. Functions are also provided for 
-    retrieving elevation data and base map tiles from the 'USGS' National Map
-    ('<https://apps.nationalmap.gov/services/>').
+    'geospatial' data, with an aim towards producing '3D' landscape 
+    visualizations in the Unity '3D' rendering engine. Functions are also 
+    provided for retrieving elevation data and base map tiles from the 'USGS' 
+    National Map ('<https://apps.nationalmap.gov/services/>').
 }
 \seealso{
 Useful links:
