+++ cran-comments.md
@@ -1,0 +1,12 @@
+## Test environments
+* local R installation, R 4.0.4
+* ubuntu 20.04 (on GitHub Actions), R devel, 4.0.4, 3.6.3
+* MacOS X 10.15.7 (on GitHub Actions), R 4.0.4
+* Windows Server 2019 (on GitHub Actions), R 4.0.4
+* win-builder (devel, 4.0.4, 3.6.3)
+
+## R CMD check results
+
+0 errors | 0 warnings | 1 note
+
+* This is a new release.
