### 2021-02-19 2b32c5b
Update NEWS
