### 2021-02-20 19d4dc9
NAIP updated; update stored tile to match
