### 2021-02-22 fca5a46
Revert "NAIP updated; update stored tile to match"
This reverts commit 19d4dc985c2357849655a274197279e4a488916d.
Apparently I spoke too soon.
