### 2021-02-22 bfa01ef
Update with comments from CRAN submission
