+++ man/terrainr-package.Rd
@@ -4,13 +4,13 @@
 \name{terrainr-package}
 \alias{terrainr}
 \alias{terrainr-package}
-\title{terrainr: Landscape Visualizations in R and Unity}
+\title{terrainr: Landscape Visualizations in R and 'Unity'}
 \description{
 \if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}
 
 Functions for the retrieval, manipulation, and visualization of 
     'geospatial' data, with an aim towards producing '3D' landscape 
-    visualizations in the Unity '3D' rendering engine. Functions are also 
+    visualizations in the 'Unity' '3D' rendering engine. Functions are also 
     provided for retrieving elevation data and base map tiles from the 'USGS' 
     National Map ('<https://apps.nationalmap.gov/services/>').
 }
