+++ man/point_from_distance.Rd
@@ -27,6 +27,9 @@ indicating the units of the provided distance.}
 \item{azimuth_unit}{A string (either \code{degrees} or \code{radians})
 indicating the units of the \code{azimuth} argument}
 }
+\value{
+An object of class [terrainr_coordinate_pair].
+}
 \description{
 Find latitude and longitude for a certain distance and azimuth from a point.
 }
