### 2021-02-22 #416 @stefaniebutland
Congratulations on passing peer review @mikemahoney218.

When you're ready, please submit a draft post by pull request, following instructions in https://blogguide.ropensci.org/. Then my colleague @steffilazerte will review it and suggest a publication date. Followup questions can move to Slack. Cheers!
